requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
blis==0.7.11
Bottleneck==1.3.2
cachetools==5.4.0
catalogue==2.0.10
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
filelock==3.16.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.33.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.24.7
idna==3.7
imageio==2.35.1
importlib_metadata==8.2.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
Levenshtein==0.25.1
libclang==18.1.1
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft==1.3.1
mkl_random==1.1.0
ml-dtypes==0.4.1
mpmath==1.3.0
murmurhash==1.0.10
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numexpr==2.8.4
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
preshed==3.0.9
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-engineio==4.9.1
python-http-client==3.3.7
python-Levenshtein==0.25.1
python-socketio==5.11.4
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.7
regex==2024.9.11
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
retry-requests==2.0.0
retrying==1.3.4
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sendgrid==6.11.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smart-open==7.0.4
spacy==3.7.6
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.5
srsly==2.4.8
starkbank-ecdsa==2.2.0
sympy==1.13.2
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.8.30
tokenizers==0.15.2
torch==2.4.1
tqdm==4.66.5
transformers==4.36.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
zipp==3.19.2
